@@ -1,6 +1,8 @@
 module github.com/teacat/ffimage
 
-go 1.19
+go 1.21
+
+toolchain go1.23.0
 
 require (
 	github.com/stretchr/testify v1.5.1
@@ -10,6 +12,7 @@ require (
 require (
 	github.com/aws/aws-sdk-go v1.44.295 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gabriel-vasile/mimetype v1.4.11 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/u2takey/ffmpeg-go v0.4.2-0.20230519105213-e1d746688945 // indirect
